requirements.txt
openai
python-dotenv
langchain_community
langchain_nvidia_ai_endpoints
faiss-cpu
streamlit
pypdf